Refactor login page by extracting to LoginPage component and enhancing form functionality
