+++ components/ui/login/LoginPage.tsx
@@ -1,9 +1,96 @@
-import React from "react";
+"use client";
+import React, { useState } from "react";
+import Image from "next/image";
 
 export default function LoginPage() {
+ 
+      
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    // Handle form submission logic here
+    console.log("Email:", email);
+    console.log("Password:", password);
+  };
+
   return (
-    <div>
-      <h1>This is a login page</h1>
+    <div className="min-h-screen flex items-center justify-center bg-[#f3f4f6]">
+      <section className="bg-white shadow-lg rounded-lg p-8 max-w-md w-full">
+        <div className="text-center mb-8">
+          <h2 className="font-bold text-2xl text-gray-800">
+            Sign In to your account
+          </h2>
+          <p className="text-gray-600">Enter your details below</p>
+        </div>
+
+        <form onSubmit={handleSubmit}>
+          <div className="mb-6">
+            <label htmlFor="email" className="block text-gray-700 mb-2">
+              Email
+            </label>
+            <input
+              id="email"
+              type="email"
+              placeholder="examplemail@gmail.com"
+              className="w-full p-3 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 transition duration-200"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+            />
+          </div>
+
+          <div className="mb-6">
+            <label htmlFor="password" className="block text-gray-700 mb-2">
+              Password
+            </label>
+            <input
+              id="password"
+              type="password"
+              placeholder="Enter your password"
+              className="w-full p-3 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 transition duration-200"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+            />
+          </div>
+
+          <button
+            type="submit"
+            className="w-full bg-textdark4 text-[#fff] py-3 rounded-lg hover:bg-blue-700 transition duration-200"
+          >
+            Sign In
+          </button>
+        </form>
+
+        <div className="text-center mt-4">
+          <a href="#" className="text-textdark3 text-base hover:text-blue-600">
+            Forgot your password?
+          </a>
+        </div>
+
+        <div className="text-center mt-4">
+          <button className="w-full bg-[#F9FAFB] text-[#606882] py-4 border border-gray-300 rounded-lg hover:text-dark hover:bg-gray-200 flex items-center justify-center transition duration-200">
+            <Image
+              src="/google.svg"
+              width={20}
+              height={20}
+              alt="Google logo"
+              className="mr-2"
+            />
+            Sign In with Google
+          </button>
+        </div>
+
+        <div className="text-center mt-4">
+          <p className="text-[#606882]">
+            Don&apos;t have an account?{" "}
+            <a href="#" className="text-textdark4 hover:text-blue-600">
+              Sign Up Now!
+            </a>
+          </p>
+        </div>
+      </section>
     </div>
   );
+
 }
